+++ kern/proc_fork.c
@@ -110,7 +110,7 @@ static int clone_oth_regions(struct proc_info * new_proc,
      * it directly that way because shared regions can't properly point to more
      * than one page table struct.
      */
-    for (size_t i = MM_HEAP_REGION; i < old_proc->mm.nr_regions; i++) {
+    for (int i = MM_HEAP_REGION; i < old_proc->mm.nr_regions; i++) {
         vm_reg_tmp = (*old_proc->mm.regions)[i];
         if (!vm_reg_tmp || (vm_reg_tmp->b_flags & B_NOTSHARED))
             continue;
@@ -252,7 +252,7 @@ pid_t proc_fork(pid_t pid)
     }
     new_proc->files->count = nofile_max;
     /* Copy and ref old file descriptors */
-    for (size_t i = 0; i < old_proc->files->count; i++) {
+    for (int i = 0; i < old_proc->files->count; i++) {
         new_proc->files->fd[i] = old_proc->files->fd[i];
         fs_fildes_ref(new_proc->files, i, 1); /* null pointer safe */
     }
